requirements.txt
python-dotenv
PyYAML 
requests 
fastapi 
uvicorn 